@@ -1,6 +1,9 @@
+from __future__ import annotations
+
+from collections.abc import Mapping, Sequence
 from dataclasses import InitVar, asdict, dataclass, field, fields
 from pathlib import Path
-from typing import ClassVar, List, Mapping, Optional, Sequence, Union
+from typing import Any, ClassVar
 
 import yaml
 
@@ -24,7 +27,7 @@ class MetaRaw:
         The type of pin data stored. This is used to determine how to read / write it.
     """
 
-    file: "str | Sequence[str] | None"
+    file: str | Sequence[str] | None
     type: str
     name: str
 
@@ -59,20 +62,20 @@ class Meta:
         A dictionary of additional metadata that may be specified by the user.
     local:
         A dictionary of additional metadata that may be added by the board, depending
-        on the backend used. E.g. RStudio Connect content id, url, etc..
+        on the backend used. E.g. Posit Connect content id, url, etc..
 
     """
 
-    _excluded: ClassVar["set[str]"] = {"name", "version", "local"}
+    _excluded: ClassVar[set[str]] = {"name", "version", "local"}
 
-    title: Optional[str]
-    description: Optional[str]
+    title: str | None
+    description: str | None
 
     # TODO(defer): different from R pins, which has a local field
     created: str
     pin_hash: str
 
-    file: Union[str, Sequence[str]]
+    file: str | Sequence[str]
     file_size: int
     type: str
 
@@ -84,14 +87,14 @@ class Meta:
     # pin_hash, created, etc.."
     version: VersionRaw
 
-    tags: Optional[List[str]] = None
-    name: Optional[str] = None
+    tags: list[str] | None = None
+    name: str | None = None
     user: Mapping = field(default_factory=dict)
     local: Mapping = field(default_factory=dict)
 
-    unknown_fields: InitVar["dict | None"] = None
+    unknown_fields: InitVar[dict | None] = None
 
-    def __post_init__(self, unknown_fields: "dict | None"):
+    def __post_init__(self, unknown_fields: dict | None):
         unknown_fields = {} if unknown_fields is None else unknown_fields
 
         self._unknown_fields = unknown_fields
@@ -102,7 +105,7 @@ class Meta:
         except KeyError:
             raise AttributeError(f"No metadata field not found: {k}")
 
-    def to_dict(self) -> Mapping:
+    def to_dict(self) -> dict[str, Any]:
         data = asdict(self)
 
         return data
@@ -119,7 +122,7 @@ class Meta:
         return d
 
     @classmethod
-    def from_pin_dict(cls, data, pin_name, version, local=None) -> "Meta":
+    def from_pin_dict(cls, data, pin_name, version, local=None) -> Meta:
         # TODO: re-arrange Meta argument positions to reflect what's been
         # learned about default arguments. e.g. title was not used at some
         # point in api_version 1
@@ -142,7 +145,7 @@ class Meta:
             unknown_fields=unknown,
         )
 
-    def to_pin_yaml(self, f: Optional[IOBase] = None) -> "str | None":
+    def to_pin_yaml(self, f: IOBase | None = None) -> str | None:
         data = self.to_pin_dict()
 
         return yaml.dump(data, f)
@@ -150,9 +153,9 @@ class Meta:
 
 @dataclass
 class MetaV0:
-    file: Union[str, Sequence[str]]
+    file: str | Sequence[str]
     type: str
-    description: "str | None"
+    description: str | None
 
     name: str
 
@@ -173,7 +176,7 @@ class MetaV0:
         return asdict(self)
 
     @classmethod
-    def from_pin_dict(cls, data, pin_name, version, local=None) -> "MetaV0":
+    def from_pin_dict(cls, data, pin_name, version, local=None) -> MetaV0:
         # could infer from dataclasses.fields(), but seems excessive.
         req_fields = {"type", "description"}
 
@@ -205,13 +208,13 @@ class MetaFactory:
 
     def get_version_for_meta(self, api_version) -> Version:
         if api_version != 1:
-            raise NotImplementedError("Unsupported api_version: %s" % api_version)
+            raise NotImplementedError(f"Unsupported api_version: {api_version}")
 
         return Version
 
     def create(
         self,
-        base_folder: "str | Path",
+        base_folder: str | Path,
         files: Sequence[StrOrFile],
         type,
         # TODO: when files is a string name should be okay as None
@@ -243,7 +246,12 @@ class MetaFactory:
 
             raise NotImplementedError("Cannot create from file object.")
         else:
-            raise NotImplementedError("TODO: creating meta from multiple files")
+            if isinstance(files, (list, tuple)):
+                from pathlib import Path
+
+                file_name = [Path(f).name for f in files]
+                file_size = [Path(f).stat().st_size for f in files]
+                version = Version.from_files(files, created)
 
         return Meta(
             title=title,
@@ -266,7 +274,7 @@ class MetaFactory:
         self,
         f: IOBase,
         pin_name: str,
-        version: "str | VersionRaw",
+        version: str | VersionRaw,
         local=None,
     ) -> Meta:
         if isinstance(version, str):
