@@ -1,5 +1,5 @@
+from collections.abc import Sequence
 from pathlib import Path
-from typing import Sequence
 
 from .config import PINS_ENV_INSECURE_READ, get_allow_pickle_read
 from .errors import PinsInsecureReadError
@@ -10,7 +10,7 @@ from .meta import Meta
 
 
 UNSAFE_TYPES = frozenset(["joblib"])
-REQUIRES_SINGLE_FILE = frozenset(["csv", "joblib", "file"])
+REQUIRES_SINGLE_FILE = frozenset(["csv", "joblib"])
 
 
 def _assert_is_pandas_df(x, file_type: str) -> None:
@@ -32,32 +32,24 @@ def _assert_is_geopandas_df(x):
         )
 
 
-def load_path(meta, path_to_version):
-    # Check that only a single file name was given
-    fnames = [meta.file] if isinstance(meta.file, str) else meta.file
-    if len(fnames) > 1 and type in REQUIRES_SINGLE_FILE:
-        raise ValueError("Cannot load data when more than 1 file")
-
+def load_path(filename: str, path_to_version, pin_type=None):
     # file path creation ------------------------------------------------------
-
-    if type == "table":  # noqa: E721 False Positive due to bug: https://github.com/rstudio/pins-python/issues/266
+    if pin_type == "table":
         # this type contains an rds and csv files named data.{ext}, so we match
         # R pins behavior and hardcode the name
-        target_fname = "data.csv"
-    else:
-        target_fname = fnames[0]
+        filename = "data.csv"
 
     if path_to_version is not None:
-        path_to_file = f"{path_to_version}/{target_fname}"
+        path_to_file = f"{path_to_version}/{filename}"
     else:
         # BoardUrl doesn't have versions, and the file is the full url
-        path_to_file = target_fname
+        path_to_file = filename
 
     return path_to_file
 
 
-def load_file(meta: Meta, fs, path_to_version):
-    return fs.open(load_path(meta, path_to_version))
+def load_file(filename: str, fs, path_to_version, pin_type):
+    return fs.open(load_path(filename, path_to_version, pin_type))
 
 
 def load_data(
@@ -88,7 +80,7 @@ def load_data(
             "  * https://scikit-learn.org/stable/modules/model_persistence.html#security-maintainability-limitations"
         )
 
-    with load_file(meta, fs, path_to_version) as f:
+    with load_file(meta.file, fs, path_to_version, meta.type) as f:
         if meta.type == "csv":
             import pandas as pd
 
@@ -119,8 +111,7 @@ def load_data(
                 import geopandas as gpd
             except ModuleNotFoundError:
                 raise ModuleNotFoundError(
-                    'The "geopandas" package is required to read "geoparquet" type '
-                    "files."
+                    'The "geopandas" package is required to read "geoparquet" type files.'
                 ) from None
 
             return gpd.read_parquet(f)
@@ -143,7 +134,7 @@ def load_data(
 
         elif meta.type == "rds":
             try:
-                import rdata
+                import rdata  # pyright: ignore[reportMissingImports]
 
                 return rdata.read_rds(f)
             except ModuleNotFoundError:
@@ -154,7 +145,9 @@ def load_data(
     raise NotImplementedError(f"No driver for type {meta.type}")
 
 
-def save_data(obj, fname, type=None, apply_suffix: bool = True) -> "str | Sequence[str]":
+def save_data(
+    obj, fname, pin_type=None, apply_suffix: bool = True
+) -> "str | Sequence[str]":
     # TODO: extensible saving with deferred importing
     # TODO: how to encode arguments to saving / loading drivers?
     #       e.g. pandas index options
@@ -163,66 +156,75 @@ def save_data(obj, fname, type=None, apply_suffix: bool = True) -> "str | Sequen
     #       of saving / loading objects different ways.
 
     if apply_suffix:
-        if type == "file":
+        if pin_type == "file":
             suffix = "".join(Path(obj).suffixes)
-        elif type == "geoparquet":
+        elif pin_type == "geoparquet":
             suffix = ".parquet"
         else:
-            suffix = f".{type}"
+            suffix = f".{pin_type}"
     else:
         suffix = ""
 
-    final_name = f"{fname}{suffix}"
+    if isinstance(fname, list):
+        final_name = fname
+    else:
+        final_name = f"{fname}{suffix}"
 
-    if type == "csv":
+    if pin_type == "csv":
         _assert_is_pandas_df(obj, file_type=type)
 
         obj.to_csv(final_name, index=False)
 
-    elif type == "arrow":
+    elif pin_type == "arrow":
         # NOTE: R pins accepts the type arrow, and saves it as feather.
         #       we allow reading this type, but raise an error for writing.
         _assert_is_pandas_df(obj, file_type=type)
 
         obj.to_feather(final_name)
 
-    elif type == "feather":
+    elif pin_type == "feather":
         _assert_is_pandas_df(obj, file_type=type)
 
         raise NotImplementedError(
             'Saving data as type "feather" no longer supported. Use type "arrow" instead.'
         )
 
-    elif type == "parquet":
+    elif pin_type == "parquet":
         _assert_is_pandas_df(obj, file_type=type)
 
         obj.to_parquet(final_name)
 
-    elif type == "geoparquet":
+    elif pin_type == "geoparquet":
         _assert_is_geopandas_df(obj)
 
         obj.to_parquet(final_name)
 
-    elif type == "joblib":
+    elif pin_type == "joblib":
         import joblib
 
         joblib.dump(obj, final_name)
 
-    elif type == "json":
+    elif pin_type == "json":
         import json
 
         json.dump(obj, open(final_name, "w"))
 
-    elif type == "file":
+    elif pin_type == "file":
         import contextlib
         import shutil
 
+        if isinstance(obj, list):
+            for file, final in zip(obj, final_name):
+                with contextlib.suppress(shutil.SameFileError):
+                    shutil.copyfile(str(file), final)
+            return obj
         # ignore the case where the source is the same as the target
-        with contextlib.suppress(shutil.SameFileError):
-            shutil.copyfile(str(obj), final_name)
+        else:
+            with contextlib.suppress(shutil.SameFileError):
+                shutil.copyfile(str(obj), final_name)
 
     else:
-        raise NotImplementedError(f"Cannot save type: {type}")
+        raise NotImplementedError(f"Cannot save type: {pin_type}")
 
     return final_name
 
