@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from pathlib import Path
 
 import fsspec
@@ -6,7 +8,7 @@ import pandas as pd
 import pytest
 
 from pins.config import PINS_ENV_INSECURE_READ
-from pins.drivers import default_title, load_data, save_data
+from pins.drivers import default_title, load_data, load_path, save_data
 from pins.errors import PinsInsecureReadError
 from pins.meta import MetaRaw
 from pins.tests.helpers import rm_env
@@ -185,3 +187,26 @@ def test_driver_apply_suffix_false(tmp_path: Path):
     res_fname = save_data(df, p_obj, type_, apply_suffix=False)
 
     assert Path(res_fname).name == "some_df"
+
+
+class TestLoadFile:
+    def test_str_file(self):
+        class _MockMetaStrFile:
+            file: str = "a"
+            type: str = "csv"
+
+        assert load_path(_MockMetaStrFile().file, None, _MockMetaStrFile().type) == "a"
+
+    def test_table(self):
+        class _MockMetaTable:
+            file: str = "a"
+            type: str = "table"
+
+        assert load_path(_MockMetaTable().file, None, _MockMetaTable().type) == "data.csv"
+
+    def test_version(self):
+        class _MockMetaTable:
+            file: str = "a"
+            type: str = "csv"
+
+        assert load_path(_MockMetaTable().file, "v1", _MockMetaTable().type) == "v1/a"
